WIP: Move Script and ScriptBuf to primitives
We want `Script` and `ScriptBuf` to live in `primitives` as plain old
data. However this is non-trivial.
