+++ internals/src/lib.rs
@@ -35,3 +35,39 @@ mod parse;
 #[cfg(feature = "serde")]
 #[macro_use]
 pub mod serde;
+
+/// Reads a `usize` from an iterator of bytes.
+// TODO: Where should this live?
+pub fn read_uint_iter(
+    data: &mut core::slice::Iter<'_, u8>,
+    size: usize,
+) -> Result<usize, UintError> {
+    if data.len() < size {
+        Err(UintError::EarlyEndOfScript)
+    } else if size > usize::from(u16::MAX / 8) {
+        // Casting to u32 would overflow
+        Err(UintError::NumericOverflow) // FIXME: Add another error variant for here?
+    } else {
+        let mut ret = 0;
+        for (i, item) in data.take(size).enumerate() {
+            ret = usize::from(*item)
+                // Casting is safe because we checked above to not repeat the same check in a loop
+                .checked_shl((i * 8) as u32)
+                .ok_or(UintError::NumericOverflow)? // FIXME: This is related to `size` not overflow.
+                .checked_add(ret)
+                // FIXME: Isn't this unreachable because we are adding a byte at a time?
+                .ok_or(UintError::NumericOverflow)?;
+        }
+        Ok(ret)
+    }
+}
+
+/// Error returned by `read_uint_iter`.
+pub enum UintError {
+    /// FIXME: This name does not make sense anymore.
+    EarlyEndOfScript,
+    /// Input data caused `usize` to overflow.
+    NumericOverflow,
+}
+
+crate::impl_from_infallible!(UintError);
