+++ primitives/Cargo.toml
@@ -16,8 +16,8 @@ exclude = ["tests", "contrib"]
 
 [features]
 default = ["std"]
-std = ["alloc", "internals/std", "io/std", "units/std"]
-alloc = ["internals/alloc", "io/alloc", "units/alloc"]
+std = ["alloc", "hex?/std", "internals/std", "io/std", "units/std"]
+alloc = ["hex?/alloc", "internals/alloc", "io/alloc", "units/alloc"]
 serde = ["dep:serde", "internals/serde", "units/serde"]
 
 [dependencies]
@@ -25,6 +25,7 @@ internals = { package = "bitcoin-internals", version = "0.3.0" }
 io = { package = "bitcoin-io", version = "0.1.1", default-features = false }
 units = { package = "bitcoin-units", version = "0.1.0", default-features = false }
 
+hex = { package = "hex-conservative", version = "0.2.0", default-features = false, optional = true }
 ordered = { version = "0.2.0", optional = true }
 
 serde = { version = "1.0.103", default-features = false, features = ["derive"], optional = true }
